feat!: Remove solve from PWG trait & introduce separate solvers for each blackbox
